+++ server/src/index.js
@@ -46,7 +46,6 @@ const PORT = process.env.PORT || 3500;
 // Environment variables
 const NODE_ENV = process.env.NODE_ENV || 'development';
 const BASE_URL = process.env.BASE_URL || '';
-const _BASE_DOMAIN = process.env.BASE_DOMAIN || 'http://localhost';
 
 // Middleware
 app.use(helmet());